\name{scSEMundir-package}
\alias{scSEMundir-package}
\alias{scSEMundir}
\docType{package}
\title{
\packageTitle{scSEMundir}
}
\description{
\packageDescription{scSEMundir}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scSEMundir}
\packageIndices{scSEMundir}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{scSEMundir}

Maintainer: \packageMaintainer{scSEMundir}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
